헤더 파일 수정 및 새로운 이미지 추가

+++ truck/views/company_views.py
@@ -1,15 +1,16 @@
-﻿from datetime import datetime
+﻿import functools
+from sqlalchemy import func
+from datetime import datetime
 from datetime import date
-
-from flask import Blueprint, render_template, request, url_for, g, flash
+from flask import Blueprint, url_for, render_template, flash, request, redirect, session, g
 from werkzeug.utils import redirect
-from flask_wtf.csrf import CSRFProtect
-
-from .. import db
+from truck import db
 from truck.models import Company, User
 from truck.forms import CompanyForm, CompanyPeriodForm, DeleteForm, CompanyClassPeriodForm
-
 from truck.views.auth_views import login_required
+
+from flask_wtf.csrf import CSRFProtect
+import logging
 
 COMPANY_CLASS_CHOICES = {
     'R': '매출처',
@@ -25,23 +26,28 @@ def create():
     form = CompanyForm()
     now_today = datetime.today().date()  # 오늘 날짜를 가져옵니다.
     today = datetime.today()
+    user_id = session['user_id']  # 로그인된 사용자 ID
+
+    if not user_id:
+        flash("User is not logged in.")
+        return redirect(url_for('auth.login'))
 
     if request.method == 'POST' and form.validate_on_submit():
         company = Company(register_date=form.register_date.data, company_class=form.company_class.data,
             business_no=form.business_no.data, firm=form.firm.data, president=form.president.data,
                 zip=form.zip.data, address=form.address.data, business_status=form.business_status.data,
                     business_item=form.business_item.data, email=form.email.data, cellphone=form.cellphone.data,
-                        tel=form.tel.data, fax=form.fax.data,  note=form.note.data)
+                        tel=form.tel.data, fax=form.fax.data,  note=form.note.data, user_id=user_id)
 
         db.session.add(company)
         db.session.commit()
         flash(f'Company table {form.firm.data} added successfully.')
-        # flash('Company Table added successfully.')
+
         return redirect(url_for('company.create', today=today))
 
     # 오늘 날짜로 입력된 모든 transport 데이터를 가져옵니다.
     companies = Company.query.filter(
-        Company.register_date == now_today
+        Company.register_date == now_today, Company.user_id == user_id
     ).all()
 
     return render_template('company/company_form.html', form=form, today=today,
@@ -49,9 +55,16 @@ def create():
 
 
 @bp.route('/company_period', methods=['GET', 'POST'])
+@login_required
 def company_period():
     form = CompanyPeriodForm()
     companies = []
+    user_id = session.get('user_id')  # 현재 로그인된 사용자 ID 가져오기
+
+    if not user_id:
+        flash("User is not logged in.")
+        return redirect(url_for('auth.login'))
+
     if request.method == 'POST' and form.validate_on_submit():
         start_date = request. form['start_date']
         end_date = request.form['end_date']
@@ -62,7 +75,7 @@ def company_period():
                 end_date = datetime.strptime(end_date, '%Y-%m-%d')
 
                 companies = Company.query.filter(
-                    Company.register_date.between(start_date, end_date)
+                    Company.register_date.between(start_date, end_date), Company.user_id == user_id
                 ).order_by(Company.register_date.desc()).all()
 
                 if not companies:
@@ -85,6 +98,11 @@ def company_period():
 def modify_company(company_id):
     company = Company.query.get_or_404(company_id)
     form = CompanyForm(obj=company)  # Company 객체로부터 초기값 설정
+    user_id = session.get('user_id')  # 현재 로그인된 사용자 ID 가져오기
+
+    if not user_id:
+        flash("User is not logged in.")
+        return redirect(url_for('auth.login'))
 
     if form.validate_on_submit():
         form.populate_obj(company)  # 폼 데이터를 Company 객체에 반영
@@ -98,35 +116,60 @@ def modify_company(company_id):
 @bp.route('/delete_company/<int:company_id>', methods=['GET', 'POST'])
 @login_required
 def delete_company(company_id):
+    user_id = session.get('user_id')  # 현재 로그인된 사용자 ID 가져오기
+
+    if not user_id:
+        flash("User is not logged in.")
+        return redirect(url_for('auth.login'))
+
+    # 삭제할 비용 항목 가져오기
     company = Company.query.get_or_404(company_id)
+
+    # 사용자가 해당 Company 항목의 소유자인지 확인
+    if company.user_id != user_id:
+        flash("You do not have permission to delete this cost entry.", 'danger')
+        return redirect(url_for('company.company_period'))
+
+    # 삭제 확인 폼
     form = DeleteForm()
+
+    # POST 요청 및 폼 유효성 검사
     if request.method == 'POST':
         db.session.delete(company)
         db.session.commit()
-        flash('Company deleted successfully.')
+        flash('Company entry deleted successfully.', 'success')
         return redirect(url_for('company.company_period'))
+
+    # 삭제 확인 페이지 렌더링
     return render_template('company/company_delete.html', form=form, company=company)
 
 
 @bp.route('/company_class_period', methods=['GET', 'POST'])
 @login_required
 def company_class_period():
-    form = CompanyClassPeriodForm()  # 들여쓰기를 수정합니다.
+    form = CompanyClassPeriodForm(company_class='P')  # 들여쓰기를 수정합니다.
     companies = None
+    user_id = session.get('user_id')  # 현재 로그인된 사용자 ID 가져오기
+
+    if not user_id:
+        flash("User is not logged in.")
+        return redirect(url_for('auth.login'))
 
     if request.method == 'POST' and form.validate_on_submit():
-        start_date = request.form['start_date']
-        end_date = request.form['end_date']
+        company_class = form.company_class.data
+        start_date = form.start_date.data
+        end_date = form.end_date.data
 
         if start_date and end_date:
             try:
-                start_date = datetime.strptime(start_date, '%Y-%m-%d')
-                end_date = datetime.strptime(end_date, '%Y-%m-%d')
+                start_date = datetime.combine(start_date, datetime.min.time())
+                end_date = datetime.combine(end_date, datetime.max.time())
 
                 # company 테이블에서 지정된 기간 동안의 데이터를 조회하고 역순으로 정렬
                 companies = Company.query.filter(
-                    Company.register_date >= start_date,
-                    Company.register_date <= end_date
+                    Company.company_class == company_class,
+                    Company.register_date.between(start_date, end_date),
+                    Company.user_id == user_id
                 ).order_by(Company.register_date.desc()).all()
 
                 # 만약 특정 cost_class를 필터링하려는 경우
@@ -134,9 +177,8 @@ def company_class_period():
                 if not companies:
                     flash('검색 결과가 없습니다.', 'warning')
                 return render_template('company/company_class_results.html',
-                            companies=companies, company_class_choices=COMPANY_CLASS_CHOICES,
-                                     enumerate=enumerate)            # COMPANY_CLASS_CHOICES
-
+                        companies=companies, company_class_choices=COMPANY_CLASS_CHOICES,
+                                company_class=company_class, enumerate=enumerate)
             except ValueError:
                 flash('잘못된 날짜 형식입니다.', 'danger')
         else:
